solve problem 4

+++ ProjectEuler/Problem001.cs
@@ -8,9 +8,9 @@ namespace ProjectEuler
 {
     public class Problem001
     {
-        public static int Run(int number = 1000)
+        public static long Run(long number = 1000)
         {
-            return Enumerable.Range(1, number - 1)
+            return Enumerable.Range(1, (int)number - 1)
                 .Where(n => n % 3 == 0 || n % 5 == 0)
                 .Sum();
         }
